single file script for fetching gas price data from ethgasstation.info
