+++ gas-price-fetcher/requirements.txt
@@ -1,0 +1,2 @@
+requests==2.25.1
+python-dotenv==0.15.0
